backend/requirements.txt
beautifulsoup4==4.13.4
Django==5.2.1
django-celery-beat==2.8.1
django-debug-toolbar==5.2.0
django-filter==25.1
djangorestframework==3.16.0
drf_spectacular==0.28.0
googletrans==3.1.0a0
httpcore==1.0.9
protobuf==6.31.0
pydantic==2.11.5
python-decouple==3.8
redis==5.0.2
Requests==2.32.3
tenacity==8.2.3
urllib3==2.4.0